+++ test/fixtures/classes/class-controller-w-oninit.js
@@ -2,6 +2,11 @@
 import Controller from 'sap/ui/core/Controller'
 
 export default class MyController extends Controller {
+  prop = 1;
+  prop2 = this.a.b;
+  prop3 = X.y;
+  control = this.byId("control");
+  other = getThing(this.prop)
   /**
    * @keep false
    */
@@ -13,7 +18,6 @@ export default class MyController extends Controller {
   onInit() {
     this.y = this.x + 1
   }
-  prop = 1;
   prop_func = () => true;
   get getter() {
     return 1